requirements-py3.txt
aniso8601
awscli
base58
boto
boto3
botocore
cffi
click
colorama
cryptography
docutils
enum34
Flask
#Flask-Ask
futures
hjson
idna
ipaddress
itsdangerous
Jinja2
jmespath
kappa
lambda-packages
MarkupSafe
ordered_set
placebo
pyasn1
pycparser
pymyra
pyOpenSSL
python-dateutil
python-slugify
PyYAML
requests
rsa
s3transfer
six
tqdm
troposphere
Unidecode
Werkzeug
wsgi-request-logger
zappa
slackclient
zdesk
lxml
python-intercom
simple-salesforce

